requirements.txt
altgraph==0.17
appdirs==1.4.4
argon2-cffi==20.1.0
asgiref==3.2.10
astroid==2.4.2
attrs==20.1.0
autopep8==1.5.4
backcall==0.2.0
bleach==3.1.5
certifi==2020.6.20
cffi==1.14.2
colorama==0.4.3
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.2
django-bootstrap-form==3.4
django-crispy-forms==1.10.0
django-filter==2.4.0
entrypoints==0.3
filelock==3.0.12
future==0.18.2
gunicorn==20.0.4
ipykernel==5.3.4
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.17.2
Jinja2==2.11.2
jsonschema==3.2.0
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
leaptask==2020.4.14
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.7
packaging==20.4
pandocfilters==1.4.2
parso==0.7.1
pefile==2019.4.18
pickleshare==0.7.5
Pillow==7.2.0
prometheus-client==0.8.0
prompt-toolkit==3.0.6
psycopg2==2.8.6
py2exe==0.9.2.2
pycodestyle==2.6.0
pycparser==2.20
pyglet==1.3.3
Pygments==2.6.1
pyleap==2020.4.24
pylint==2.5.3
pylint-django==2.3.0
pylint-plugin-utils==0.6
pyparsing==2.4.7
PyQt5==5.15.1
PyQt5-sip==12.8.1
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
pywinpty==0.5.7
pyzmq==19.0.2
qtconsole==4.7.6
QtPy==1.9.0
Send2Trash==1.5.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
terminado==0.8.3
testpath==0.4.4
toml==0.10.1
tornado==6.0.4
traitlets==4.3.3
typing-extensions==3.7.4.3
virtualenv==20.2.2
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.2.0
widgetsnbextension==3.5.1
wrapt==1.12.1
